Moving MVCache decls to a separate header file
Moving the MVCache class declaration to a separate C++ header
file. This makes it easy to test the class.
